+++ lib/ui/scaffolds/login.dart
@@ -3,7 +3,7 @@ import 'package:courses_in_english/model/user/user.dart';
 import 'package:courses_in_english/ui/basic_components/line_separator.dart';
 import 'package:courses_in_english/ui/basic_components/scenery_widget.dart';
 import 'package:courses_in_english/ui/scaffolds/bnb_home.dart';
-import 'package:courses_in_english/ui/scaffolds/create_user_screen.dart';
+import 'package:courses_in_english/ui/scaffolds/create_user.dart';
 import 'package:flutter/material.dart';
 
 class LoginScreen extends StatefulWidget {
